+++ src/assetbundles/resources/src/js/modules/light-switch.js
@@ -1,7 +1,8 @@
 import { createObserver } from '../util/create-observer';
 
-export default function onInit() {
-    document.querySelectorAll('[data-toggle-listener]').forEach((node) => {
+export default {};
+export function onInit(context) {
+    context.querySelectorAll('[data-toggle-listener]').forEach((node) => {
         let oldValue = node.getAttribute('aria-checked');
         const checkbox = node.querySelector('input[type="hidden"]');
         const observer = createObserver((mutation) => {